requirements.txt
diffusers==0.18.1
matplotlib==3.4.2
pandas==1.4.2
Pillow==10.0.0
protobuf==4.23.4
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
